Trove Fishing Bot
This is a Trove bot i made because i could not get 7s or any other bot
to work recuse they would get stuck or the tooltip would disapere, this
is for the ones that could not get the other bots to work in the game so
hopefully this will work for you

Note that you will need to edit this your self (
Adress = 0x106345330 ; Enter your Dynamic/static Pointer Address here.
no offsets is needed only the address for when there is a fish on your
hook and the value turns int to 1.
Baits = 96 ; Enter the amount of Baits you have here, how many baits you
have in the inv when the bot starts.
§:: ; Enter the Hotkey you want to use to start the Bot.

also note that dropping boots isn't something i use so i don't have it
in the code, you could make a simple iamagesearch/pixel search for it to
drag out all the boots you fish up if you want.

and the final note!

IF you don't like my bot then don't simply use it or write a unnecessary
comment on it, i made his bot for myself so you should be happy that i
make it open source and that i show it to you.

if you have some Questions about it then PM me or comment.

if you want help finding pointers look at the Youtube video about how to
get pointers in the tutorial section.
)